Account for missing adjustments and version aware delete and copy

+++ cdsobs/cli/_delete_dataset.py
@@ -16,7 +16,9 @@ from cdsobs.config import CDSObsConfig
 from cdsobs.observation_catalogue.database import get_session
 from cdsobs.observation_catalogue.models import Catalogue
 from cdsobs.observation_catalogue.repositories.catalogue import CatalogueRepository
-from cdsobs.observation_catalogue.repositories.dataset import CadsDatasetRepository
+from cdsobs.observation_catalogue.repositories.dataset_version import (
+    CadsDatasetVersionRepository,
+)
 from cdsobs.observation_catalogue.schemas.catalogue import (
     CatalogueSchema,
     CliCatalogueFilters,
@@ -38,14 +40,21 @@ def delete_dataset(
         help="Filter by an exact date or by an interval of two dates. For example: "
         "to delete all partitions of year 1970: 1970-1-1,1970-12-31",
     ),
+    version: str = typer.Option(None, help="Version to delete"),
+    dry_run: bool = typer.Option(
+        False, help="Do nothing but print the entries to be deleted."
+    ),
 ):
     """Permanently delete the given dataset from the catalogue and the storage."""
-    confirm = prompt(
-        "This will delete the data permanently."
-        " This action cannot be undone. "
-        "Please type again the name of the dataset to confirm"
-    )
-    assert confirm == dataset, CliException("Error: The entered value do not match.")
+    if not dry_run:
+        confirm = prompt(
+            "This will delete the data permanently."
+            " This action cannot be undone. "
+            "Please type again the name of the dataset to confirm"
+        )
+        assert confirm == dataset, CliException(
+            "Error: The entered value do not match."
+        )
     try:
         init_config = CDSObsConfig.from_yaml(cdsobs_config_yml)
     except ConfigError:
@@ -53,27 +62,37 @@ def delete_dataset(
 
     with get_session(init_config.catalogue_db) as catalogue_session:
         deleted_entries = delete_from_catalogue(
-            catalogue_session, dataset, dataset_source, time
+            catalogue_session,
+            dataset,
+            dataset_source,
+            time,
+            version=version,
+            dry_run=dry_run,
         )
-        s3_client = S3Client.from_config(init_config.s3config)
-        try:
-            delete_from_s3(deleted_entries, s3_client)
-        except (Exception, KeyboardInterrupt):
-            catalogue_rollback(catalogue_session, deleted_entries)
-            raise
-        nd = len(deleted_entries)
-        console.print(
-            f"[bold green] {nd} entries deleted from {dataset}. [/bold green]"
-        )
-        nremaining = catalogue_session.scalar(
-            select(func.count()).select_from(Catalogue)
-        )
-        if nremaining == 0:
-            CadsDatasetRepository(catalogue_session).delete_dataset(dataset)
+        if not dry_run:
+            s3_client = S3Client.from_config(init_config.s3config)
+            try:
+                delete_from_s3(deleted_entries, s3_client)
+            except (Exception, KeyboardInterrupt):
+                catalogue_rollback(catalogue_session, deleted_entries)
+                raise
+            nd = len(deleted_entries)
             console.print(
-                f"[bold green] Deleted {dataset} from datasets table as it was left empty. "
-                f"[/bold green]"
+                f"[bold green] {nd} entries deleted from {dataset}. [/bold green]"
             )
+            nremaining = catalogue_session.scalar(
+                select(func.count())
+                .select_from(Catalogue)
+                .where(Catalogue.dataset == dataset, Catalogue.version == version)
+            )
+            if nremaining == 0:
+                CadsDatasetVersionRepository(catalogue_session).delete_dataset(
+                    dataset, version
+                )
+                console.print(
+                    f"[bold green] Deleted {dataset} {version} from datasets table as it was "
+                    f"left empty. [/bold green]"
+                )
 
 
 def delete_from_catalogue(
@@ -81,8 +100,13 @@ def delete_from_catalogue(
     dataset: str,
     dataset_source: str,
     time: str,
+    version: str,
+    dry_run: bool = False,
 ):
     catalogue_repo = CatalogueRepository(catalogue_session)
+    versions = [
+        version,
+    ]
     filters = CliCatalogueFilters(
         dataset=dataset,
         dataset_source=dataset_source,
@@ -91,17 +115,23 @@ def delete_from_catalogue(
         latitudes=[],
         variables=[],
         stations=[],
-        versions=[],
+        versions=versions,
         deprecated="all",
     ).to_repository_filters()
     entries = catalogue_repo.get_by_filters(filters)
     if not len(entries):
         console.print(f"[red] No entries for dataset {dataset} found")
-    try:
-        catalogue_session.execute(delete(Catalogue).where(*filters))
-        catalogue_session.commit()
-    except (Exception, KeyboardInterrupt):
-        catalogue_rollback(catalogue_session, entries)
+    if dry_run:
+        assets = [e.asset for e in entries]
+        console.print(
+            f"Would delete {len(entries)} entries with the following assets: {assets}"
+        )
+    else:
+        try:
+            catalogue_session.execute(delete(Catalogue).where(*filters))
+            catalogue_session.commit()
+        except (Exception, KeyboardInterrupt):
+            catalogue_rollback(catalogue_session, entries)
     return entries
